Added ll_apply(...) Functionality
Added the function ll_apply(...). This function applys a function over the list
of elements. The function to be supplied needs to have the following:
	* Two arguments:
		* First argument being the element to pass to the function.
		* Second argument being an argument to pass to the function.

The first argument to the supplied function is not specified when calling
ll_apply(...) as it will be supplied by the function ll_apply(...) itself. The
second argument must be specified when calling ll_apply(...) as it is not
supplied by the function ll_apply(...) itself.

TODO:
 * Test ll_apply(...).

File(s):
   modified:   include/linkedlist.h
   modified:   src/linkedlist.c